fix(status): show delivery queue warnings in detailed reports
