+++ src/commands/status-all/diagnosis.ts
@@ -26,7 +26,11 @@ import {
   formatPluginCompatibilityNotice,
   type PluginCompatibilityNotice,
 } from "../../plugins/status.js";
-import type { StatusGatewayDiagnosticsResult } from "../status-runtime-shared.ts";
+import { formatDeliveryQueueHealthLine } from "../health-format.js";
+import type {
+  resolveStatusGatewayHealthSafe,
+  StatusGatewayDiagnosticsResult,
+} from "../status-runtime-shared.ts";
 import {
   formatUpdateRestartActionLines,
   formatUpdateRestartStatusValue,
@@ -164,7 +168,7 @@ export async function appendStatusAllDiagnosis(params: {
   exporterDiagnostics: StatusGatewayDiagnosticsResult | null;
   agentStatus?: AgentStatusLike;
   gatewayReachable: boolean;
-  health: unknown;
+  health: Awaited<ReturnType<typeof resolveStatusGatewayHealthSafe>> | null | undefined;
   nodeOnlyGateway: NodeOnlyGatewayInfo | null;
 }) {
   const { lines, muted, ok, warn, fail } = params;
@@ -501,31 +505,19 @@ export async function appendStatusAllDiagnosis(params: {
     );
   }
 
-  const healthErr = (() => {
-    if (!params.health || typeof params.health !== "object") {
-      return "";
+  if (params.health) {
+    if ("error" in params.health) {
+      if (params.health.error) {
+        lines.push("");
+        lines.push(muted("Gateway health:"));
+        lines.push(`  ${muted(redactStatusSecrets(params.health.error))}`);
+      }
+    } else {
+      const deliveryQueueLine = formatDeliveryQueueHealthLine(params.health);
+      if (deliveryQueueLine) {
+        emitCheck(redactStatusSecrets(deliveryQueueLine), "warn");
+      }
     }
-    const record = params.health as Record<string, unknown>;
-    if (!("error" in record)) {
-      return "";
-    }
-    const value = record.error;
-    if (!value) {
-      return "";
-    }
-    if (typeof value === "string") {
-      return value;
-    }
-    try {
-      return JSON.stringify(value, null, 2);
-    } catch {
-      return "[unserializable error]";
-    }
-  })();
-  if (healthErr) {
-    lines.push("");
-    lines.push(muted("Gateway health:"));
-    lines.push(`  ${muted(redactStatusSecrets(healthErr))}`);
   }
 
   lines.push("");
