+++ src/commands/status.command-sections.ts
@@ -13,6 +13,7 @@ import { formatDurationCompact } from "../infra/format-time/format-duration.js";
 import type { HeartbeatEventPayload } from "../infra/heartbeat-events.js";
 import type { Tone } from "../memory-host-sdk/status.js";
 import type { SessionStatus, StatusSummary } from "../status/types.js";
+import { formatDeliveryQueueHealthLine } from "./health-format.js";
 import type { HealthSummary } from "./health.js";
 import type { AgentLocalStatus } from "./status.agent-local.js";
 import type { MemoryStatusSnapshot, MemoryPluginStatus } from "./status.scan.shared.js";
@@ -268,7 +269,7 @@ export function buildStatusSecurityAuditLines(params: {
   return lines;
 }
 
-/** Builds health table rows from gateway health and channel health text. */
+/** Builds gateway, channel, and delivery queue health table rows. */
 export function buildStatusHealthRows(params: {
   health: HealthSummary;
   formatHealthChannelLines: (summary: HealthSummary, opts: { accountMode: "all" }) => string[];
@@ -290,7 +291,12 @@ export function buildStatusHealthRows(params: {
       Detail: formatEventLoopHealthDetail(params.health.eventLoop),
     });
   }
-  for (const line of params.formatHealthChannelLines(params.health, { accountMode: "all" })) {
+  const healthLines = params.formatHealthChannelLines(params.health, { accountMode: "all" });
+  const deliveryQueueLine = formatDeliveryQueueHealthLine(params.health);
+  if (deliveryQueueLine) {
+    healthLines.push(deliveryQueueLine);
+  }
+  for (const line of healthLines) {
     const colon = line.indexOf(":");
     if (colon === -1) {
       continue;
@@ -298,7 +304,7 @@ export function buildStatusHealthRows(params: {
     const item = line.slice(0, colon).trim();
     const detail = line.slice(colon + 1).trim();
     const normalized = normalizeLowercaseStringOrEmpty(detail);
-    // Channel health format is string-based; classify known prefixes into table status chips.
+    // Shared health text uses known prefixes to classify table status chips.
     const status =
       normalized === "healthy" || normalized.startsWith("ok") || normalized.startsWith("configured")
         ? params.ok("OK")
